backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.110.3
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==79.0.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.25.0
watchfiles==1.0.5
websockets==15.0.1
XlsxWriter==3.2.3
